on process of user menu

+++ client/src/components/ProfileMenu/ProfileMenu.jsx
@@ -1,30 +1,38 @@
 import React from 'react'
 import {Avatar, Menu} from '@mantine/core';
 import { useNavigate } from 'react-router-dom'
+import '@mantine/core/styles.css';
 
 
-export const ProfileMenu = ({user, logout}) => {
+ const ProfileMenu = ({user, logout}) => {
   const navigate = useNavigate();
+
+  if(!user){
+    return null;
+  }
+
   return (
-    <Menu>
+    <Menu trigger="hover"
+    delay={500}
+    transition="fade"
+    placement="end"
+    withinPortal>
         <Menu.Target>
-          <div role="button" aria-haspopup="menu" aria-expanded="false">
-          <Avatar src={user?.picture} alt='user image' radius={"lg"}/>
-          </div>
+        <Avatar src={user.picture || "logo.png"} alt='user image' radius={"xl"}/>
         </Menu.Target>
-        
-        <Menu.Dropdown role="menu">
-            <Menu.Item role="menuitem">
+
+        <Menu.Dropdown>
+            <Menu.Item>
                 Favourites
             </Menu.Item>
 
-            <Menu.Item role="menuitem">
+            <Menu.Item>
                 Bookings
             </Menu.Item>
 
-            <Menu.Item role="menuitem" onClick={()=>{
-              localStorage.clear();
-              logout()
+            <Menu.Item onClick={()=>{
+                localStorage.clear();
+                logout()
             }}>
                 Logout
             </Menu.Item>
